change to rapidfuzz

+++ mealie/services/parser_services/ingredient_parser.py
@@ -1,8 +1,8 @@
 from abc import ABC, abstractmethod
 from fractions import Fraction
 
-from fuzzywuzzy import fuzz
 from pydantic import UUID4
+from rapidfuzz import fuzz
 from sqlalchemy.orm import Session
 
 from mealie.core.root_logger import get_logger
@@ -78,7 +78,7 @@ class ABCIngredientParser(ABC):
 
         # further refine match using fuzzy matching
         best_match: IngredientFood | None = None
-        best_match_score: int = 0
+        best_match_score: float = 0
 
         for item in response.items:
             score = fuzz.ratio(food.name, item.name)
@@ -101,7 +101,7 @@ class ABCIngredientParser(ABC):
 
         # further refine match using fuzzy matching
         best_match: IngredientUnit | None = None
-        best_match_score: int = 0
+        best_match_score: float = 0
 
         for item in response.items:
             score = fuzz.ratio(unit.name, item.name)
